VIDEO/requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
# 基础依赖
# 固定 setuptools 版本以解决 pkg_resources 弃用警告
setuptools<81
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS>=3.0.0
# 图像识别
ultralytics==8.3.141
# PyTorch（OpenVINO 2025.3.0需要PyTorch 2.1+）
torch>=2.1.0
torchvision>=0.16.0
# ONNX模型推理支持
onnxruntime-gpu==1.20.0
# OpenVINO模型导出支持
openvino>=2024.0.0
# 图像处理
opencv-python==4.8.0.74
Pillow==10.0.0
# 工具类
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
psycopg2-binary==2.9.9
python-magic==0.4.27
numpy==1.26.4
pytz==2024.1
# 提供标准化健康检查端点
py-healthcheck==1.10.0
# 网卡探测
netifaces==0.11.0
# 进程管理
psutil==7.2.2
# ONVIF摄像头支持
WSDiscovery==2.1.2
onvif-zeep==0.2.12
# 任务调度
APScheduler==3.10.4
# Cron表达式解析
croniter>=1.4.1
# Kafka消息队列
kafka-python>=2.0.2
# OCR识别
paddlepaddle==3.2.0
paddleocr==3.2.0
# 语音转文本
pydub==0.25.1
